% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OpenLongData.R
\name{as_list}
\alias{as_list}
\title{Retrieve components from an OpenLong data set}
\usage{
as_list(x, ...)
}
\arguments{
\item{x}{an object inheriting from the \code{OpenLong::OpenLongData} class.}
}
\description{
OpenLong objects contain longitudinal and baseline components.
}
