% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OpenLongData.R
\name{data_clean}
\alias{data_clean}
\title{Clean existing variables within an OpenLong data set}
\usage{
data_clean(x, ...)
}
\arguments{
\item{x}{an object inheriting from the \code{OpenLong::OpenLongData} class.}
}
\description{
Modifies existing variables within data from the study of interest
}
