% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OpenLongData.R
\name{data_derive}
\alias{data_derive}
\title{Create new variables within an OpenLong data set}
\usage{
data_derive(x, ...)
}
\arguments{
\item{x}{an object inheriting from the \code{OpenLong::OpenLongData} class.}
}
\description{
generates new variables using data from the study of interest
}
