% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OpenLongData.R
\name{data_load}
\alias{data_load}
\title{Load files for an OpenLong data set}
\usage{
data_load(x, ...)
}
\arguments{
\item{x}{an object inheriting from the \code{OpenLong::OpenLongData} class.}
}
\description{
retrieve relevant files for the study of interest
}
