% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_abc.R
\docType{data}
\name{health_abc}
\alias{health_abc}
\title{Synthetic Health ABC data}
\format{
\subsection{\code{health_abc}}{

A data frame with X rows and Y columns:
\describe{
\item{country}{Country name}
\item{year}{Year}
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
health_abc
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
