% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenLongAric.R
\name{OpenLongAric}
\alias{OpenLongAric}
\title{An object to hold ARIC data}
\usage{
OpenLongAric(
  filepath = class_missing,
  baseline = class_missing,
  longitudinal = class_missing,
  components = class_missing,
  loaded = class_missing,
  excluded = class_missing,
  cleaned = class_missing,
  derived = class_missing
)
}
\arguments{
\item{filepath}{a path to the location of Biolincc's MESA package.
If \code{NULL} or missing, an error will be thrown. Eventually,
a \code{NULL} or missing value will imply that simulated data will
be used.}
}
\value{
An \code{OpenLong::OpenLongAric} S7 object.
}
\description{
Holds a set of baseline and longitudinal data from ARIC.
}
\seealso{
Other constructors: 
\code{\link{OpenLongMesa}()}
}
\concept{constructors}
