% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OpenLongData.R
\name{print.OpenLong::OpenLongData}
\alias{print.OpenLong::OpenLongData}
\title{Print an open long data object}
\usage{
\method{print}{`OpenLong::OpenLongData`}(x, ...)
}
\arguments{
\item{x}{an object of class \code{OpenLong::OpenLongData}}

\item{...}{not currently used}
}
\value{
\code{x}, invisibly
}
\description{
Print an open long data object
}
